requirements.txt
Django>=4.0.0,<4.1.0
djangorestframework>=3.13.0,<3.14.0
pillow>=8.4.0,<9.0.0
django-cors-headers>=3.10.1,<3.11.0
